selftests/bpf: add tc helpers
The tc_tunnel.c test defines small fonctions to easily attach eBPF
programs to tc hooks, either on egress, ingress or both.

Move those helpers in a dedicated file so that other tests can benefit
from it.
